Faster response time for Time Trade window
The code that Arick made available was good but load time was a little
long. Playing with applications like lightbox and fancybox I was able to
find a way to make the load time a little faster and fit the link to the
current design of the site (with the Time Trade button). Window load
time is much faster. Loaded up complete version of site. This doesn't
include the animation that William completed last night.