Update for website
I updated the web pages for HeroesforStudents. There is now a green
donate button for Home, WhatWeDo and WhoWeAre webpages that go to the
subscribe site for Heroes on Paypal.  Also updated the TimeTrade button
so it has a blue timetrade button. Changed the stylesheet for the
animation on the WhatWeDo page so that the screen is more centered on
the page. Also added the video from the old home page to the WhoWeAre
page.  For all pages I changed the 'Register' link  and created a link
of Heroes and for Teachers to register. When the links are clicked each
go to the original forms from the current website.  Login and Get
Involved pages are still inoperable for now.