Going Live
Added a version 5 which marks the code used to 'go live' on the
HeroesforStudents.org site. The paragraphs for the What We Do page were
re-styled so that the paragraphs are rendering as paragraphs and the
font was changed to 13pts to match the font size of the Who We Are page.
The Who We Are page had the Co-Founder information and photos added to
it (under the Vimeo video).  William's animation was added to the What
We Do page to reflect the full video.