Vimeo for What We Do animation added
I created a Vimeo account to upload William's animation to so that the
code could be placed into the What We Do page and allow rendering on all
browser platforms without incident. This will be temporary until we get
account information from Chris and so that Heroes for Students.org will
be ready for their presentation on March 4, 2015. with the deliverables
we agreed on.  Screen sizes for the videos on What We Do and Who We Are
pages were adjusted to a width of 1000 and a height of 600 to provide a
bigger screen for presentation purposes.