Version 0.6 Heroes
This update includes the videos being on the What We Do page and only
the photos and bios of the co-founders being on the Who We Are page. The
videos on the What We Do page are images and when the images are clicked
lightbox opens a window and starts playing the video automatically.